Added hitboxes

+++ src/Tools/Image.java
@@ -17,46 +17,46 @@ public class Image {
     private int x;
     private int y;
 
-    public Image(Wrap wrap, String path, int x, int y, int width, int height) {
+    public Image(Wrap wrap, String path, double x, double y, int width, int height) {
         this.wrap = wrap;
         try {
             image = ImageIO.read(new File(path));
         } catch (IOException e) {
             System.out.println("Couldn't find the file");
         }
-        this.x = (int) ((double)x * wrap.getScale());
-        this.y = (int) ((double)y * wrap.getScale());
+        this.x = (int) (x * wrap.getScale());
+        this.y = (int) (y * wrap.getScale());
         this.width = (int) ((double)width * wrap.getScale());
         this.height = (int) ((double)height * wrap.getScale());
     }
 
-    public Image(Wrap wrap, String path, int x, int y) {
+    public Image(Wrap wrap, String path, double x, double y) {
         this.wrap = wrap;
         try {
             image = ImageIO.read(new File(path));
         } catch (IOException e) {
             System.out.println("Couldn't find the file");
         }
-        this.x = (int) ((double)x * wrap.getScale());
-        this.y = (int) ((double)y * wrap.getScale());
+        this.x = (int) (x * wrap.getScale());
+        this.y = (int) (y * wrap.getScale());
         this.width = (int)((double)image.getWidth() * wrap.getScale());
         this.height = (int)((double)image.getHeight() * wrap.getScale());
     }
 
-    public Image (Wrap wrap, BufferedImage image, int x, int y, int width, int height) {
+    public Image (Wrap wrap, BufferedImage image, double x, double y, int width, int height) {
         this.wrap = wrap;
         this.image = image;
-        this.x = (int) ((double)x * wrap.getScale());
-        this.y = (int) ((double)y * wrap.getScale());
+        this.x = (int) (x * wrap.getScale());
+        this.y = (int) (y * wrap.getScale());
         this.width = (int) ((double)width * wrap.getScale());
         this.height = (int) ((double)height * wrap.getScale());
     }
 
-    public Image (Wrap wrap, BufferedImage image, int x, int y) {
+    public Image (Wrap wrap, BufferedImage image, double x, double y) {
         this.wrap = wrap;
         this.image = image;
-        this.x = (int) ((double)x * wrap.getScale());
-        this.y = (int) ((double)y * wrap.getScale());
+        this.x = (int) (x * wrap.getScale());
+        this.y = (int) (y * wrap.getScale());
         this.width = (int)((double)image.getWidth() * wrap.getScale());
         this.height = (int)((double)image.getHeight() * wrap.getScale());
     }
@@ -73,6 +73,7 @@ public class Image {
             e.printStackTrace();
         }
     }
+
     public void changeImage(BufferedImage image){
         this.image = image;
     }
